feat: split up some code
